# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Outdoor Event Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/comments:
    get:
      operationId: listEventComments
      tags: [comments]
      summary: List comments for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Create a comment on an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /admin/participants/search:
    get:
      operationId: adminSearchParticipants
      tags: [admin]
      summary: Quick participant search by name, phone, or email
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Matching participants (max 20)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Participant"
  /admin/participants:
    get:
      operationId: adminListParticipants
      tags: [admin]
      summary: Admin paginated participant list with search
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Participant list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminParticipantListResponse"
    post:
      operationId: adminCreateParticipant
      tags: [admin]
      summary: Admin create a participant manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminParticipantInput"
      responses:
        "201":
          description: Created participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
...
  /admin/participants/{phone}:
    get:
      operationId: adminGetParticipant
      tags: [admin]
      summary: Get participant details with registration history
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Participant with registrations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminParticipantDetail"
        "404":
          description: Not found
    patch:
      operationId: adminUpdateParticipant
      tags: [admin]
      summary: Update participant info
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminParticipantInput"
      responses:
        "200":
          description: Updated participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
  /admin/participants/{phone}/phone:
    patch:
      operationId: adminChangeParticipantPhone
      tags: [admin]
      summary: Change a participant's phone number
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePhoneInput"
      responses:
        "200":
          description: Updated participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Participant"
        "409":
          description: Phone already in use
  /admin/participants/{phone}/registrations:
    post:
      operationId: adminAddParticipantToEvent
      tags: [admin]
      summary: Admin add a participant to an event
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRegistrationInput"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /admin/events/{id}/export:
    get:
      operationId: adminExportEventRegistrations
      tags: [admin]
      summary: Export event registrations as CSV
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /admin/sponsors/{id}/export:
    get:
      operationId: adminExportSponsorReport
      tags: [admin]
      summary: Export sponsor analytics report as CSV
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
components:
  schemas:
...
    Registration:
      type: object
      required: [id, eventId, name, email, status, registeredAt, referralCount]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        name:
          type: string
        fullNameAr:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        phoneCountryCode:
          type: string
        nationality:
          type: ["string", "null"]
        nationalityOther:
          type: ["string", "null"]
        hasMedicalConditions:
          type: boolean
        medicalDetails:
          type: ["string", "null"]
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        waiverAcceptedAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, waitlist, cancelled]
        referralToken:
          type: ["string", "null"]
        referredByRegistrationId:
          type: ["integer", "null"]
        referralCount:
          type: integer
        registeredAt:
          type: string
          format: date-time
...
    Comment:
      type: object
      required: [id, eventId, authorName, content, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [authorName, content]
      properties:
        authorName:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
...
    Participant:
      type: object
      required: [id, email, name, totalPoints, totalEvents, referralCount, joinedAt]
      properties:
        id:
          type: integer
        phone:
          type: ["string", "null"]
        email:
          type: string
        name:
          type: string
        totalPoints:
          type: integer
        totalEvents:
          type: integer
        referralCount:
          type: integer
        emergencyContactName:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        waiverAcceptedAt:
          type: ["string", "null"]
          format: date-time
        joinedAt:
          type: string
          format: date-time
    AdminParticipantInput:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        emergencyContactName:
          type: string
        emergencyContactPhone:
          type: string
        totalPoints:
          type: integer
    AdminRegistrationInput:
      type: object
      required: [eventId]
      properties:
        eventId:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, waitlist, cancelled]
    AdminParticipantListResponse:
      type: object
      required: [participants, total, page, limit]
      properties:
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminParticipantDetail:
      type: object
      required: [participant, registrations]
      properties:
        participant:
          $ref: "#/components/schemas/Participant"
        registrations:
          type: array
          items:
            $ref: "#/components/schemas/AdminRegistrationSummary"
    AdminRegistrationSummary:
      type: object
      required: [id, eventId, eventTitle, eventDate, status, registeredAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        eventTitle:
          type: string
        eventDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, confirmed, waitlist, cancelled]
        registeredAt:
          type: string
          format: date-time
...
    ChangePhoneInput:
      type: object
      required: [newPhone]
      properties:
        newPhone:
          type: string
...